Display merge default when `$request` is empty
When the `$_REQUEST` super global has a key set without a value the
merge shortcode returns that empty value. Generally this is not
desired.

This is especially prevalant with `$_GET` variables coming from an
autoresponder email. `http://example.com/page/?Name=%$name$%` and more
is hardcoded in. When 1ShoppingCart fills the merge on their end,
those fields merge nothing if the value doesn't exist for the customer.

When that is the case, the specified default text should be displayed.

[Addresses: GH-7]